Remove deposit option for smart margin
