+++ pages/market.tsx
@@ -21,7 +21,7 @@ import TradeIsolatedMargin from 'sections/futures/Trade/TradeIsolatedMargin';
 import TransferIsolatedMarginModal from 'sections/futures/Trade/TransferIsolatedMarginModal';
 import DelayedOrderConfirmationModal from 'sections/futures/TradeConfirmation/DelayedOrderConfirmationModal';
 import TradeConfirmationModalCrossMargin from 'sections/futures/TradeConfirmation/TradeConfirmationModalCrossMargin';
-import DepositWithdrawCrossMargin from 'sections/futures/TradeCrossMargin/DepositWithdrawCrossMargin';
+import WithdrawSmartMargin from 'sections/futures/TradeCrossMargin/WithdrawSmartMargin';
 import AppLayout from 'sections/shared/Layout/AppLayout';
 import { setOpenModal } from 'state/app/reducer';
 import { selectShowModal, selectShowPositionModal } from 'state/app/selectors';
@@ -120,11 +120,8 @@ const Market: MarketComponent = () => {
 					onDismiss={() => dispatch(setOpenModal(null))}
 				/>
 			)}
-			{(openModal === 'futures_cross_deposit' || openModal === 'futures_cross_withdraw') && (
-				<DepositWithdrawCrossMargin
-					defaultTab="deposit"
-					onDismiss={() => dispatch(setOpenModal(null))}
-				/>
+			{openModal === 'futures_cross_withdraw' && (
+				<WithdrawSmartMargin onDismiss={() => dispatch(setOpenModal(null))} />
 			)}
 
 			{openModal === 'futures_confirm_smart_margin_trade' && <TradeConfirmationModalCrossMargin />}
